Refactor ComponentItemModal, updates and new utility
Refactored ComponentItemModal component to handle more reusable behaviours and migrated specific functionalities out of the old component to keep it clean and maintainable. ComponentItemAction was updated to accommodate these changes. Utility function 'copyOnMouseEvent' was added to be used for text copying events to reduce redundant code blocks for the same use-case.
