+++ src/utils/copy-on-mouse-event.ts
@@ -1,0 +1,9 @@
+import {MouseEvent} from "react";
+
+export function copyOnMouseEvent<TMouseEvent = unknown>(text: string): (e: MouseEvent<TMouseEvent>) => void {
+    return function $onMouseEvent(e: MouseEvent<TMouseEvent>) {
+        e.stopPropagation();
+
+        window.navigator.clipboard.writeText(text).then();
+    };
+}
